beta ft_strtrim :rofl:

+++ ft_strrchr.c
@@ -1,7 +1,7 @@
 #include "libft.h"
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t l = ft_strlen(s);
+	long long l = ft_strlen(s);
 
 	while(l >= 0)
 	{
